# oas/openAPI_remediated.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "42C Bank API",
        "description": "An intentionally unsecure banking API for education and training purposes.",
        "contact": {
            "name": "42Crunch Support",
            "url": "https://support.42crunch.com",
            "email": "support@42crunch.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://{domain}/apis/banking/{version}",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "https"
                },
                "domain": {
                    "default": "localhost:3000",
                    "description": "local development"
                },
                "version": {
                    "enum": [
                        "v1",
                        "v2"
                    ],
                    "default": "v1"
                }
            }
        }
    ],
    "security": [
        {
            "AccessToken": []
        }
    ],
...
    "paths": {
...
        "/account": {
            "get": {
                "description": "Get a user account",
                "operationId": "GetAccount",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "accountId": {
                                            "$ref": "#/components/schemas/AccountId"
                                        },
                                        "name": {
                                            "$ref": "#/components/schemas/Name"
                                        },
                                        "email": {
                                            "$ref": "#/components/schemas/Email"
                                        },
                                        "address": {
                                            "$ref": "#/components/schemas/Address"
                                        }
                                    },
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            },
            "delete": {
                "description": "Delete an Account",
                "operationId": "DeleteAccount",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageObject"
                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            }
        },
...
        "/account/products/cards": {
            "post": {
                "description": "Apply for a credit card",
                "operationId": "CreateCardApplication",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "delivery": {
                                        "$ref": "#/components/schemas/DeliveryType"
                                    }
                                },
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "$ref": "#/components/schemas/ResponseMessage"
                                        },
                                        "referenceId": {
                                            "type": "number"
                                        }
                                    },
                                    "additionalProperties": false
                                }
                            }
                        },
                        "headers": {
                            "X-Rate-Limit-Limit": {
                                "description": "The number of allowed requests in the current period",
                                "schema": {
                                    "type": "integer",
                                    "minimum": 0,
                                    "maximum": 100,
                                    "format": "int32"                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            },
            "get": {
                "description": "Get credit card application",
                "operationId": "GetCardApplication",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id": {
                                            "type": "number"
                                        },
                                        "name": {
                                            "$ref": "#/components/schemas/Name"
                                        },
                                        "delivery": {
                                            "$ref": "#/components/schemas/DeliveryType"
                                        },
                                        "address": {
                                            "$ref": "#/components/schemas/Address"
                                        },
                                        "status": {
                                            "$ref": "#/components/schemas/CardStatus"
                                        }
                                    },
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            }
        },
        "/account/products/cards/{referenceId}": {
            "put": {
                "description": "Update the delivery method and address",
                "operationId": "UpdateCardApplication",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/referenceId"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "delivery": {
                                        "$ref": "#/components/schemas/DeliveryType"
                                    },
                                    "postalAddress": {
                                        "$ref": "#/components/schemas/Address"
                                    }
                                },
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id": {
                                            "type": "number"
                                        },
                                        "name": {
                                            "$ref": "#/components/schemas/Name"
                                        },
                                        "delivery": {
                                            "$ref": "#/components/schemas/DeliveryType"
                                        },
                                        "address": {
                                            "$ref": "#/components/schemas/Address"
                                        },
                                        "status": {
                                            "$ref": "#/components/schemas/CardStatus"
                                        }
                                    },
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            },
            "delete": {
                "description": "Cancel a credit card application",
                "operationId": "DeleteCardApplication",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/referenceId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "$ref": "#/components/schemas/ResponseMessage"
                                        }
                                    },
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            }
        },
        "/account/payees": {
            "get": {
                "description": "Get list of payees",
                "operationId": "GetAccountPayees",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "maxItems": 10,
                                    "items": {
                                        "oneOf": [
                                            {
                                                "$ref": "#/components/schemas/UtilityPayee"
                                            },
                                            {
                                                "$ref": "#/components/schemas/ContactPayee"
                                            }
                                        ],
                                        "discriminator": {
                                            "propertyName": "payeeType"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            },
            "post": {
                "description": "Add new payee",
                "operationId": "CreatePayee",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/UtilityPayee"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ContactPayee"
                                    }
                                ],
                                "discriminator": {
                                    "propertyName": "payeeType"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type":"object",
                                    "properties":{
                                        "message":{
                                            "$ref": "#/components/schemas/ResponseMessage"
                                        },
                                        "payeeId":{
                                            "$ref": "#/components/schemas/AccountId"
                                        }
                                    },
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            }
        },
        "/account/payees/{payeeId}": {
            "delete": {
                "description": "Delete a payee from the payee list",
                "operationId": "DeletePayee",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/payeeId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageObject"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401ErrorResponse"
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            }
        },
        "/account/payments": {
            "post": {
                "description": "Create a payment to make a transfer or pay a bill.",
                "operationId": "CreatePayment",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "payeeId": {
                                        "$ref": "#/components/schemas/AccountId"
                                    },
                                    "amount": {
                                        "$ref": "#/components/schemas/TransferAmount"
                                    },
                                    "currency": {
                                        "$ref": "#/components/schemas/Currency"
                                    }
                                },
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "transactionId": {
                                            "$ref": "#/components/schemas/TransactionId"
                                        },
                                        "message": {
                                            "$ref": "#/components/schemas/ResponseMessage"
                                        }
                                    },
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "406": {
                        "$ref": "#/components/responses/406ErrorResponse"
                    },
                    "415": {
                        "$ref": "#/components/responses/415ErrorResponse"
                    },
                    "429": {
                        "$ref": "#/components/responses/429ErrorResponse"
                    },
                    "default": {
                        "$ref": "#/components/responses/DefaultErrorResponse"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AccountId": {
                "description": "A unique and immutable identifier used to identify the account resource. This identifier has no meaning to the account owner.",
                "type": "string",
                "pattern": "^[a-f\\d]{24}$",
                "example": "661cc8464787094021349548"
            },
...
            "Email": {
                "type": "string",
                "format": "email-secured",
                "example": "jbernard@email.com",
                "pattern": "^\\w+(?:[\\-.]\\w+)*@[0-9A-Za-z]+(?:[\\-.][0-9A-Za-z]+)*\\.(?:[A-Za-z][0-9A-Za-z]+|[0-9]+[A-Za-z][0-9A-Za-z]*)$",
                "minLength": 6,
                "maxLength": 320,
                "x-42c-format": "o:email-secured"
            },
            "Name": {
                "type": "string",
                "description": "persons name",
                "pattern": "^[A-Za-z0-9. -]{1,50}$",
                "example": "Jean Bernard"
            },
...
        },
        "responses": {
...
            "406ErrorResponse": {
                "description": "Not Acceptable",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MessageObject"
                        }
                    }
                }
            },
            "415ErrorResponse": {
                "description": "Unsupported Media Type",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MessageObject"
                        }
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "AccessToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}